Fixed return value in get_inode_index.

+++ simplefs/simplefs_utils.c
@@ -125,7 +125,7 @@ int32_t next_inode(uint16_t prev_inode, char* name, uint8_t type, void* shm_addr
 
 
 int32_t get_inode_index(char *path, uint8_t type, void* shm_addr){
-    uint16_t current_inode = 1;
+    int32_t current_inode = 1;
 
     char** sub_path = NULL;
     uint32_t sub_path_count = parse_path(path, &sub_path);
